+++ Source/Server/Game/Network/NetworkSend.cs
@@ -682,7 +682,7 @@ namespace Server
                                             {
                                                 {
                                                     ref var withBlock2 = ref Core.Type.Map[mapNum].Event[i].Pages[X].CommandList[Y].Commands[z];
-                                                    buffer.WriteByte(withBlock2.Index);
+                                                    buffer.WriteInt32(withBlock2.Index);
                                                     buffer.WriteString(withBlock2.Text1);
                                                     buffer.WriteString(withBlock2.Text2);
                                                     buffer.WriteString(withBlock2.Text3);
